+++ src/utils/manifest.js
@@ -4,7 +4,7 @@ export default function getManifest() {
   if (isBookmarklet()) {
     return {
       version_name: '0.5.0-beta',
-      description: 'A collection of modules that improve your Gann OnCampus experience',
+      description: 'A collection of modules that improve your Gann MyGann experience',
     };
   } else {
     return chrome.runtime.getManifest();